pyproject.toml
[tool.poetry]
name = "personal_site"
version = "0.1.0"
description = ""
authors = ["Corentin <corentin.zheng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
whitenoise = "^6.1.0"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
